+++ src/main/kotlin/fi/hsl/jore4/hastus/service/exporting/validation/LastStopNotTimingPointException.kt
@@ -1,4 +1,4 @@
-package fi.hsl.jore4.hastus.export.validation
+package fi.hsl.jore4.hastus.service.exporting.validation
 
 import org.springframework.http.HttpStatus
 import org.springframework.web.server.ResponseStatusException
